+++ jamanchu/src/main/java/com/recipe/jamanchu/auth/LoginFilter.java
@@ -2,7 +2,7 @@ package com.recipe.jamanchu.auth;
 
 import com.recipe.jamanchu.entity.UserEntity;
 import com.recipe.jamanchu.exceptions.exception.UserNotFoundException;
-import com.recipe.jamanchu.model.dto.request.UserDetailsDTO;
+import com.recipe.jamanchu.model.dto.request.users.UserDetailsDTO;
 import com.recipe.jamanchu.repository.UserRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.Cookie;
