+++ jamanchu/src/main/java/com/recipe/jamanchu/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.recipe.jamanchu.controller;
 
-import com.recipe.jamanchu.model.dto.request.SignupDTO;
-import com.recipe.jamanchu.model.dto.response.UserResponse;
+import com.recipe.jamanchu.model.dto.request.users.SignupDTO;
+import com.recipe.jamanchu.model.type.ResultCode;
 import com.recipe.jamanchu.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +17,7 @@ public class UserController {
   private final UserService userService;
 
   @PostMapping("/signup")
-  public ResponseEntity<UserResponse> signup(SignupDTO signupDTO) {
+  public ResponseEntity<ResultCode> signup(SignupDTO signupDTO) {
 
     return ResponseEntity.ok(userService.signup(signupDTO));
   }
