+++ jamanchu/src/main/java/com/recipe/jamanchu/model/type/ResultCode.java
@@ -1,4 +1,4 @@
-package com.recipe.jamanchu.model.dto.response;
+package com.recipe.jamanchu.model.type;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public enum UserResponse {
+public enum ResultCode {
 
   SUCCESS_SIGNUP(HttpStatus.CREATED, "회원가입 성공!");
