+++ jamanchu/src/main/java/com/recipe/jamanchu/service/UserService.java
@@ -1,10 +1,10 @@
 package com.recipe.jamanchu.service;
 
-import com.recipe.jamanchu.model.dto.request.SignupDTO;
-import com.recipe.jamanchu.model.dto.response.UserResponse;
+import com.recipe.jamanchu.model.dto.request.users.SignupDTO;
+import com.recipe.jamanchu.model.type.ResultCode;
 
 public interface UserService {
 
-  UserResponse signup(SignupDTO signupDTO);
+  ResultCode signup(SignupDTO signupDTO);
 
 }
